+++ construct_typed/generic_wrapper.py
@@ -14,6 +14,10 @@ if t.TYPE_CHECKING:
     from construct import Array as Array
     from construct import Construct as Construct
     from construct import ListContainer as ListContainer
+    from construct import Subconstruct as Subconstruct
+    from construct import SymmetricAdapter as SymmetricAdapter
+    from construct import Tunnel as Tunnel
+    from construct import Validator as Validator
     from construct.core import ConstantOrContextLambda as ConstantOrContextLambda
     from construct.core import Context as Context
     from construct.core import PathType as PathType
@@ -22,26 +26,50 @@ if t.TYPE_CHECKING:
 else:
     import construct as cs
 
-    # at runtime, the original classes are no generics, so whe have to make new classes with generics support
+    # at runtime, the original classes are not generic, so we have to make new classes with generics support
     class Construct(t.Generic[ParsedType, BuildTypes], cs.Construct):
-        pass
+        """Subscriptable version of `construct.Construct` that can be used as a generic type."""
+
+    class Subconstruct(
+        t.Generic[SubconParsedType, SubconBuildTypes, ParsedType, BuildTypes],
+        cs.Subconstruct,
+    ):
+        """Subscriptable version of `construct.Subconstruct` that can be used as a generic type."""
 
     class Adapter(
         t.Generic[SubconParsedType, SubconBuildTypes, ParsedType, BuildTypes],
         cs.Adapter,
     ):
-        pass
+        """Subscriptable version of `construct.Adapter` that can be used as a generic type."""
 
-    class ListContainer(t.Generic[ListType], cs.ListContainer):
-        pass
+    class Tunnel(
+        t.Generic[SubconParsedType, SubconBuildTypes],
+        cs.Tunnel,
+    ):
+        """Subscriptable version of `construct.Tunnel` that can be used as a generic type."""
 
-    class Context:
-        pass
+    class SymmetricAdapter(
+        t.Generic[SubconParsedType, SubconBuildTypes, ParsedType, BuildTypes],
+        cs.SymmetricAdapter,
+    ):
+        """Subscriptable version of `construct.SymmetricAdapter` that can be used as a generic type."""
+
+    class Validator(
+        t.Generic[SubconParsedType, SubconBuildTypes],
+        cs.Validator,
+    ):
+        """Subscriptable version of `construct.Validator` that can be used as a generic type."""
 
     class Array(
         t.Generic[SubconParsedType, SubconBuildTypes],
         cs.Array,
     ):
+        """Subscriptable version of `construct.Array` that can be used as a generic type."""
+
+    class ListContainer(t.Generic[ListType], cs.ListContainer):
+        pass
+
+    class Context:
         pass
 
     ConstantOrContextLambda = ValueType | t.Callable[[Context], t.Any]